refactor(lib::track): move duration formatting to own module
