+++ lib/src/track/mod.rs
@@ -1,6 +1,8 @@
+mod fmt_ext;
 mod full_track;
 
+pub use fmt_ext::DurationFmtShort;
 pub use full_track::{
-    DurationFmtShort, LyricData, MediaTypes, MediaTypesSimple, MetadataOptions, PodcastTrackData,
-    RadioTrackData, Track, TrackData, TrackMetadata, parse_metadata_from_file,
+    LyricData, MediaTypes, MediaTypesSimple, MetadataOptions, PodcastTrackData, RadioTrackData,
+    Track, TrackData, TrackMetadata, parse_metadata_from_file,
 };
